cat: EXAddressSingle, and a component bound that follows the form
The FT-991A's chart prints ONE menu number per row — "P1 : 001 - 153" — not
a group/subgroup/item triple, so its EX address field is three digits and
its printed number IS the whole address. This package had two forms and one
component ceiling of 99, which made rows 100 to 153 not merely unregistered
but INEXPRESSIBLE: V8 refused them, and under uint8 a generated literal of
300 failed at compile time in a file no human wrote.

EXAddressSingle joins Triple and Pair as a third EXAddressForm.
wireEXAddress gains a "%03d" arm and stays the only place an address becomes
wire digits, so Dialect.EXAddressWidth() — which MEASURES that render —
answers 3 and exReadLen, exAnswerMinLen and exAnswerMaxLen follow with no
second table of widths anywhere: 6, 7 and, for a chart whose widest item is
8 digits, 14. ParseEXAddress gains a three-digit branch keyed (P1,0,0) via a
threeDigitsAt beside twoDigitsAt, with its own refusal sentences; the
existing "six" and "four" spellings are untouched, and the doc comment now
says what the golden actually pins — ONE sentence of the four, not four —
so that a later reader knows the other five rest on a test alone.

V12 gains the Single clause: P2 and P3 must both be zero, because the render
drops both and a component silently dropped from every frame is what that
rule exists to make impossible. It is the Pair rule one component further
down, and its refusal names the offending index and the address as the frame
would have carried it.

EXAddress's components widen from uint8 to uint16, which is what makes the
STATED bound the operative one. Under uint8 the TYPE was the bound at 255 —
a number matching no document, neither the field's 999 nor any chart's row
count — and any Single clause would have been vacuous whatever it said. V8's
bound is now form-dependent: Triple and Pair keep 99 and their SHIPPED
sentence byte for byte, while Single's P1 gets 999 and a sentence of its own
naming the form, the bound and the render width. maxEXComponent's own doc
comment is amended with them, not just its use: it stated the whole
rationale in %02d-overrun terms and claimed uint8 constrained the value,
both of which are now true of two thirds of the rule.

The nine cross-check and inventory tests that named uint8 for an address
component follow the type. No expected VALUE moves with them:
evidence-literals.golden is keyed by literal, and a type name is not one.

Byte identity: `go generate ./core/cat/...` reproduces the four inventories
byte-identically (their literals are untyped), and the twenty frozen paths
diff empty, evidence-literals.golden included.

singleDialect joins allTestDialects() for the reason pairDialect did — every
gate, conformance and round-trip property this package states over "any
dialect" is now stated over a three-digit one, and all of them pass.

+++ core/cat/dialectequiv_test.go
@@ -291,7 +291,7 @@ func TestNewDialect_DerivedEXIndicesMatchRecomputation(t *testing.T) {
 	for p1 := 0; p1 <= 9; p1++ {
 		for p2 := 0; p2 <= 20; p2++ {
 			for p3 := 0; p3 <= 25; p3++ {
-				a := EXAddress{P1: uint8(p1), P2: uint8(p2), P3: uint8(p3)}
+				a := EXAddress{P1: uint16(p1), P2: uint16(p2), P3: uint16(p3)}
 				want := inInventory[a]
 				swept++
 				if want {
